extract types and add account check
